test: split inventory import export routes
